requirements.txt
numpy==2.3.4
pygame==2.6.1
Pillow>=10.0.0
